requirements.txt
psutil
pywin32; sys_platform == 'win32'
cryptography